Implemented Full-text Search
If the query string has multiple keywords, each should be treated as a prefix. For example, the query "good u" is first tokenized into two tokens: "good" and "u". The search results will be all the movie titles that contain a word starting with "good" and a word starting with "u", such as "The Good, The Bad and The Ugly", and "Goodbye Uncle Tom".